add account endpoints, start refactor for different data source
Going to refactor this design to fit around the design of Banner's student API. The student API is already organized similar to this API, and allows us to get directly at the data without dealing with any SQL. However, I'm going to refactor the work-study out as well as remove the dual enrollment hours from the academic status endpoint. We have a specific query we need to use for work-study, and I think that work-study would be better for a different API since it deals with highly confidential data. For dual enrollment hours, we don't really have a good way to get this from banner 9's REST student api since it's a OSU specific function. If this data is still needed, I'm going to look into putting this into its own api as well.