change dockerfile to be from different context
I was wanting to copy the swagger.yaml file to the docker image, but it wouldn't allow me to copy files in the parent directory of the build context. The fix was to run 'docker build' from the root directory of the repository so that it's able to get swagger.yaml in the build context.